backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
pytesseract==0.3.13
easyocr==1.7.2
Pillow==10.4.0
httpx==0.27.0
pydantic==2.9.0
pydantic-settings==2.5.0
python-dotenv==1.0.1
numpy
pyspellchecker==0.8.1
